factory refactoring
