+++ src/main/java/UI/LoginUI.java
@@ -1,12 +1,7 @@
 package UI;
 
-import config.CommonCryptograph;
-import config.Cryptograph;
-import entity_layer.CommonUserFactory;
-import config.UserStorage;
 import use_cases.login.LoginController;
 import use_cases.login.LoginException;
-import use_cases.user_registration.UserRegistrationInteractor;
 
 import javax.swing.*;
 import java.awt.event.*;
