+++ src/main/java/entity_factories/CommonUserFactory.java
@@ -1,4 +1,8 @@
-package entity_layer;
+package entity_factories;
+
+import entity_factories.CommonUser;
+import entity_layer.User;
+import entity_layer.UserFactory;
 
 public class CommonUserFactory implements UserFactory {
