+++ src/test/java/LoginTests.java
@@ -2,10 +2,8 @@ import config.CommonCryptograph;
 import config.Cryptograph;
 import config.UserDataStoreGateway;
 import config.UserStorage;
-import entity_layer.CommonUser;
-import entity_layer.CommonUserFactory;
+import entity_factories.CommonUser;
 import entity_layer.User;
-import entity_layer.UserFactory;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeAll;
 import use_cases.login.LoginController;
